Make filterMediaUsing() available in SpatieMediaLibraryImageColumn and SpatieMediaLibraryImageEntry
