+++ packages/spatie-laravel-media-library-plugin/README.md
@@ -220,6 +220,28 @@ SpatieMediaLibraryImageColumn::make('avatar')
     ->conversion('thumb')
 ```
 
+### Filtering media
+
+It's possible to target a column component to only display a certain subset of media in a collection. To do that, you
+can filter the media collection using the `filterMediaUsing()` method. This method accepts a function that receives the
+`$media` collection and manipulates it. You can use
+any [collection method](https://laravel.com/docs/collections#available-methods) to filter it.
+
+For example, you could scope the field to only display media that has certain custom properties:
+
+```php
+use Filament\Tables\Columns\SpatieMediaLibraryImageColumn;
+use Illuminate\Support\Collection;
+
+SpatieMediaLibraryImageColumn::make('images')
+    ->filterMediaUsing(
+        fn (Collection $media): Collection => $media->where(
+            'custom_properties.gallery_id',
+            12345
+        ),
+    )
+```
+
 ## Infolist entry
 
 To use the media library image entry:
@@ -263,4 +285,26 @@ use Filament\Infolists\Components\SpatieMediaLibraryImageEntry;
 
 SpatieMediaLibraryImageEntry::make('avatar')
     ->conversion('thumb')
+```
+
+### Filtering media
+
+It's possible to target an infolist component to only display a certain subset of media in a collection. To do that, you
+can filter the media collection using the `filterMediaUsing()` method. This method accepts a function that receives the
+`$media` collection and manipulates it. You can use
+any [collection method](https://laravel.com/docs/collections#available-methods) to filter it.
+
+For example, you could scope the field to only display media that has certain custom properties:
+
+```php
+use Filament\Tables\Columns\SpatieMediaLibraryImageEntry;
+use Illuminate\Support\Collection;
+
+SpatieMediaLibraryImageEntry::make('images')
+    ->filterMediaUsing(
+        fn (Collection $media): Collection => $media->where(
+            'custom_properties.gallery_id',
+            12345
+        ),
+    )
 ```
