+++ packages/spatie-laravel-media-library-plugin/src/Resources/Concerns/HasMediaFilter.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace Filament\Resources\Concerns;
+
+use Closure;
+use Illuminate\Support\Collection;
+
+trait HasMediaFilter
+{
+    protected ?Closure $filterMediaUsing = null;
+
+    public function filterMediaUsing(?Closure $callback): static
+    {
+        $this->filterMediaUsing = $callback;
+
+        return $this;
+    }
+
+    public function filterMedia(Collection $media): Collection
+    {
+        return $this->evaluate($this->filterMediaUsing, [
+            'media' => $media,
+        ]) ?? $media;
+    }
+
+    public function hasMediaFilter(): bool
+    {
+        return $this->filterMediaUsing instanceof Closure;
+    }
+}
